Moved the application controller under relative controllers directory
