+++ apps/legal-gazette-api/src/modules/applications/application.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { SequelizeModule } from '@nestjs/sequelize'
 
 import { CaseModel } from '../case/case.model'
-import { ApplicationController } from './application.controller'
+import { ApplicationController } from './controllers/application.controller'
 
 @Module({
   imports: [SequelizeModule.forFeature([CaseModel])],
